# Conflicts:
#   src/components/styled/styled-card/index.js

+++ src/components/defaultPage/menu/index.jsx
@@ -1,16 +1,17 @@
 import React, { useState } from 'react';
 import StyledMenu from '../../styled/styled-menu';
-import { MenuCenter, MenuLeft, MenuRight, StyledLogo } from './styled';
+import { MenuCenter, MenuLeft, MenuRight, StyledLogo, StyledUser, NameUser } from './styled';
 import { Grid, Feed, Dropdown } from 'semantic-ui-react';
 import { AiOutlineHome, AiFillHome } from 'react-icons/ai';
 import { BsSearch } from 'react-icons/bs';
 import { RiSearchFill } from 'react-icons/ri';
 import UserDefault from '../../../assets/img/userDefault.jpg';
 import LogoMenu from '../../../assets/img/LogoBrancoVerde.png';
-import { Link, useHistory } from 'react-router-dom';
+import { useHistory, useLocation } from 'react-router-dom';
 import { login } from '../../../redux/actions/session';
+import Swal from 'sweetalert2';
 
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 import './menu.css';
 
@@ -18,8 +19,11 @@ const TopBar = () => {
   const [activeHome, setActiveHome] = useState(true);
   const [activeSearch, setActiveSearch] = useState(false);
   const dispatch = useDispatch();
+  const session = useSelector((state) => state.session);
 
+  console.log(session);
   const history = useHistory();
+  const location = useLocation();
 
   return (
     <>
@@ -33,35 +37,51 @@ const TopBar = () => {
             </Grid.Column>
             <Grid.Column>
               <MenuCenter>
-                {!activeHome ? (
-                  <div
-                    className="div-home"
-                    onClick={() => {
-                      setActiveHome(!activeHome);
-                      setActiveSearch(false);
-                      history.push('/');
-                    }}>
-                    <AiOutlineHome className="icon-home" />
-                  </div>
-                ) : (
-                  <div className="div-home-active">
-                    <AiFillHome className="icon-home" />
-                  </div>
+                {location.pathname == '/' && (
+                  <>
+                    <div className="div-home-active">
+                      <AiFillHome className="icon-home" />
+                    </div>
+                    <div
+                      className="div-search"
+                      onClick={() => {
+                        history.push('/home');
+                      }}>
+                      <BsSearch className="icon-search" />
+                    </div>
+                  </>
                 )}
-                {!activeSearch ? (
-                  <div
-                    className="div-search"
-                    onClick={() => {
-                      setActiveSearch(!activeSearch);
-                      setActiveHome(false);
-                      history.push('/home');
-                    }}>
-                    <BsSearch className="icon-search" />
-                  </div>
-                ) : (
-                  <div className="div-search-active">
-                    <RiSearchFill className="icon-search" />
-                  </div>
+                {location.pathname == '/home' && (
+                  <>
+                    <div
+                      className="div-home"
+                      onClick={() => {
+                        history.push('/');
+                      }}>
+                      <AiOutlineHome className="icon-home" />
+                    </div>
+                    <div className="div-search-active">
+                      <RiSearchFill className="icon-search" />
+                    </div>
+                  </>
+                )}
+                {location.pathname === '/shelves' && (
+                  <>
+                    <div
+                      className="div-home"
+                      onClick={() => {
+                        history.push('/');
+                      }}>
+                      <AiOutlineHome className="icon-home" />
+                    </div>
+                    <div
+                      className="div-search"
+                      onClick={() => {
+                        history.push('/home');
+                      }}>
+                      <BsSearch className="icon-search" />
+                    </div>
+                  </>
                 )}
               </MenuCenter>
             </Grid.Column>
@@ -70,11 +90,15 @@ const TopBar = () => {
                 <Feed>
                   <Feed.Event>
                     <Feed.Label className="user-default">
-                      <img src={UserDefault} />
+                      {session && session.user.image_url ? (
+                        <StyledUser src={session.user.image_url} />
+                      ) : (
+                        <img src={UserDefault} />
+                      )}
                     </Feed.Label>
                   </Feed.Event>
                 </Feed>
-                <Dropdown direction="left" text={<b> Bruno </b>}>
+                <Dropdown direction="left" text={<NameUser>{session.user.name} </NameUser>}>
                   <Dropdown.Menu>
                     <Dropdown.Item
                       icon="user"
@@ -87,8 +111,15 @@ const TopBar = () => {
                       color="red"
                       text="Sair"
                       onClick={() => {
-                        window.localStorage.clear();
-                        dispatch(login('', ''));
+                        Swal.fire({
+                          title: `Volte logo, ${session.user.name}!`,
+                          confirmButtonText: `Sair`,
+                        }).then((result) => {
+                          if (result.isConfirmed) {
+                            window.localStorage.clear();
+                            dispatch(login('', ''));
+                          }
+                        });
                       }}
                     />
                   </Dropdown.Menu>
